Retrieve data from BacDive

Retrieve data from BacDive

retrieve_data(searchTerm, searchType = "taxon")

Arguments

searchTerm
Mandatory character string (in case of searchType = sequence, culturecollectionno or taxon) or integer (in case of bacdive_id), specifying what shall be searched for.

searchType
Mandatory character string that specifies which type of search will be performed (technically, which API endpoint). Can be taxon (default), bacdive_id, sequence, or culturecollectionno.

Value

A list of lists containing either a single BacDive dataset in case the searchTerm was unambiguous (bacdive_id, sequence, culturecollectionno), or a large list containing all datasets that match an ambiguous searchTerm (most taxons).

Examples

  

dataset_717 <- retrieve_data(searchTerm = 717, searchType = "bacdive_id")

#> 717 

  

dataset_DSM_319 <- retrieve_data(searchTerm = "DSM 319", searchType = "culturecollectionno")

#> 20319 

  

dataset_AJ000733 <- retrieve_data(searchTerm = "AJ000733", searchType = "sequence")

#> 000733 

  

dataset_Bss <- retrieve_data(searchTerm = "Bacillus subtilis subtilis")

#>  

#> Warning: 131 BacDive datasets are being downloaded... Please note: Depending on your internet connection speed, and the BacDive server load right now, this might take some time and may cause R(Studio) to become temporarily unresponsive. Maybe go for a walk for a few (I'm guessing: 0.5 to 0.7) minutes ;-)
#> 

#> 2 

#> Data download in progress for BacDive-IDs: 

#> 1156 

#> 1157 

#> 1158 

#> 1159 

#> 1160 

#> 1161 

#> 1162 

#> 1163 

#> 1164 

#> 1165 

#> 1166 

#> 1167 

#> 1168 

#> 1169 

#> 1170 

#> 1171 

#> 1172 

#> 1173 

#> 1174 

#> 1175 

#> 1176 

#> 1177 

#> 1178 

#> 1179 

#> 1180 

#> 1181 

#> 1182 

#> 1183 

#> 1184 

#> 1185 

#> 1186 

#> 134612 

#> 134701 

#> 134853 

#> 134854 

#> 134856 

#> 134857 

#> 134858 

#> 135155 

#> 135173 

#> 135255 

#> 135269 

#> 135270 

#> 135280 

#> 135297 

#> 135365 

#> 135545 

#> 135547 

#> 135577 

#> 135581 

#> 135590 

#> 135620 

#> 135643 

#> 135676 

#> 135695 

#> 135696 

#> 135697 

#> 135698 

#> 135699 

#> 135700 

#> 135701 

#> 135783 

#> 135784 

#> 135909 

#> 136058 

#> 136375 

#> 136391 

#> 136392 

#> 136394 

#> 136395 

#> 136501 

#> 136948 

#> 136959 

#> 136969 

#> 136972 

#> 136975 

#> 136991 

#> 136992 

#> 137087 

#> 137100 

#> 137129 

#> 137348 

#> 137349 

#> 137416 

#> 137419 

#> 137433 

#> 137923 

#> 138480 

#> 138711 

#> 138730 

#> 138763 

#> 138773 

#> 138777 

#> 138779 

#> 138875 

#> 138915 

#> 138956 

#> 138957 

#> 138958 

#> 138964 

#> 138965 

#> 138967 

#> 138968 

#> 138970 

#> 138971 

#> 138972 

#> 138973 

#> 138974 

#> 138975 

#> 138976 

#> 138979 

#> 138980 

#> 138981 

#> 138982 

#> 138983 

#> 138984 

#> 138985 

#> 138986 

#> 138987 

#> 138988 

#> 138989 

#> 138990 

#> 138992 

#> 138993 

#> 138994 

#> 138995 

#> 139209 

#> 139407 

#> 139411 

#> 139520 

#> 139538