Retrieve Results of an Advanced Search/Query

Retrieve Results of an Advanced Search/Query

retrieve_search_results(queryURL)

Arguments

queryURL
URL of a query prepared on BacDive.DSMZ.de/AdvSearch, as explained in the vignette "Pre-Configuring Advanced Searches and Retrieving the Results".

Value

A list of lists containing either a single BacDive dataset in case the searchTerm was unambiguous (bacdive_id, sequence, culturecollectionno), or a large list containing all datasets that match an ambiguous searchTerm (most taxons).

Examples

plant_human_pathogens <- retrieve_search_results(queryURL = "https://bacdive.dsmz.de/advsearch?advsearch=search&site=advsearch&searchparams%5B5%5D%5Bsearchterm%5D=1&searchparams%5B156%5D%5Bsearchterm%5D=&searchparams%5B158%5D%5Bsearchterm%5D=1")

#> Downloading 155 datasets will probably take 38 to 78 seconds.

#> Downloading BacDive IDs 1905 to 133015 (but not necessarily contiguously): 

#> 1905 

#> 1906 

#> 1911 

#> 1917 

#> 1919 

#> 1935 

#> 2083 

#> 2086 

#> 2891 

#> 2893 

#> 2900 

#> 2901 

#> 2902 

#> 4306 

#> 4307 

#> 4308 

#> 4309 

#> 4346 

#> 4349 

#> 4350 

#> 4351 

#> 4352 

#> 4353 

#> 4369 

#> 4371 

#> 4379 

#> 4384 

#> 4386 

#> 4388 

#> 4389 

#> 4390 

#> 4393 

#> 4396 

#> 4397 

#> 4398 

#> 4402 

#> 4404 

#> 4997 

#> 5001 

#> 5002 

#> 5007 

#> 5008 

#> 5009 

#> 5010 

#> 5011 

#> 5017 

#> 5018 

#> 5022 

#> 5025 

#> 5026 

#> 5027 

#> 5028 

#> 5131 

#> 5180 

#> 5249 

#> 5251 

#> 7286 

#> 7289 

#> 7290 

#> 7294 

#> 7299 

#> 7338 

#> 7482 

#> 7484 

#> 7485 

#> 10721 

#> 10883 

#> 11315 

#> 11356 

#> 11357 

#> 12808 

#> 12817 

#> 12863 

#> 12927 

#> 13016 

#> 13024 

#> 13044 

#> 13051 

#> 13052 

#> 13053 

#> 13058 

#> 13059 

#> 13060 

#> 13101 

#> 13102 

#> 13103 

#> 13104 

#> 13107 

#> 13140 

#> 13143 

#> 13144 

#> 13167 

#> 13168 

#> 13571 

#> 13573 

#> 13647 

#> 14199 

#> 14222 

#> 14230 

#> 14353 

#> 14356 

#> 14359 

#> 14883 

#> 15146 

#> 15586 

#> 16229 

#> 16315 

#> 16371 

#> 16372 

#> 16373 

#> 16387 

#> 17584 

#> 17589 

#> 17591 

#> 17592 

#> 17593 

#> 17599 

#> 17600 

#> 17602 

#> 17605 

#> 17606 

#> 17607 

#> 17608 

#> 17609 

#> 17610 

#> 17611 

#> 17612 

#> 17613 

#> 17614 

#> 17615 

#> 17616 

#> 17621 

#> 17622 

#> 17954 

#> 22995 

#> 23003 

#> 23005 

#> 23013 

#> 23936 

#> 24401 

#> 24420 

#> 24487 

#> 24622 

#> 130234 

#> 130238 

#> 130317 

#> 130574 

#> 130598 

#> 131277 

#> 131278 

#> 132107 

#> 132585 

#> 132902 

#> 132918 

#> 133015