Retrieve Results of an Advanced Search/Query

Retrieve Results of an Advanced Search/Query

retrieve_search_results(queryURL)

Arguments

queryURL
URL of a query prepated on BacDive.DSMZ.de/AdvSearch, as explained in the vignette "The semi-automatic approach: retrieving data through a pre-configured advanced search".

Value

A list of lists containing either a single BacDive dataset in case the searchTerm was unambiguous (bacdive_id, sequence, culturecollectionno), or a large list containing all datasets that match an ambiguous searchTerm (most taxons).

Examples

retrieve_search_results(queryURL = "https://bacdive.dsmz.de/advsearch?site=advsearch&searchparams[78][contenttype]=text&searchparams[78][typecontent]=contains&searchparams[78][searchterm]=Miller&advsearch=search")

#> Data download in progress for BacDive-IDs: 

#> 132598 

#> 133012 

#> $`132598`
#> $`132598`$taxonomy_name
#> $`132598`$taxonomy_name$strains_tax_PNU
#>   species_epithet subspecies_epithet is_type_strain   domain       phylum
#> 1         mayonii                 NA          FALSE Bacteria Spirochaetes
#>          class ordo       family status_fam    genus  status_gen
#> 1 Spirochaetes   NA Borreliaceae         NA Borrelia genus (AL) 
#>            species   status_spec               full_scientific_name
#> 1 Borrelia mayonii sp. nov. (VP) Borrelia mayonii Pritt et al. 2016
#>   ID_reference
#> 1        20215
#> 
#> $`132598`$taxonomy_name$strains
#>     domain       phylum        class           ordo       family    genus
#> 1 Bacteria Spirochaetes Spirochaetes Spirochaetales Borreliaceae Borrelia
#>            species species_epithet subspecies_epithet
#> 1 Borrelia mayonii         mayonii                 NA
#>                 full_scientific_name designation is_type_strain ID_reference
#> 1 Borrelia mayonii Pritt et al. 2016   MN14-1420          FALSE        24387
#> 
#> 
#> $`132598`$morphology_physiology
#> named list()
#> 
#> $`132598`$culture_growth_condition
#> $`132598`$culture_growth_condition$culture_medium
#>                                           medium_name medium_growth
#> 1 BSK-MEDIUM (DSMZ Medium 403), 30°C, microaerophilic          TRUE
#>                                                         media_link
#> 1 https://www.dsmz.de/microorganisms/medium/pdf/DSMZ_Medium403.pdf
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          medium_composition
#> 1 <br><B>Solution A:</B><br> CMRL 1066/10 x Glutamin (GIBCO 042-1545) 100.0 ml<br> Proteose peptone No 2 (DIFCO 0121-01-3) 5.0 g<br> Proteose tryptone (DIFCO 0123-01) 1.0 g<br> Proteose yeastolate (TC) (DIFCO 5577-15) 1.0 g<br> HEPES buffer (acid) (SIGMA H3375) 6.0 g<br> Na-citrate 0.7 g<br> Glucose (alpha-, D+) 3.0 g<br> Na-pyruvate 0.8 g<br> N-Acetylglucosamine 0.4 g<br> MgCl<SUB>2</SUB> x 6 H<SUB>2</SUB>O 0.3 g<br> L-Glutamine 0.1 g<br> Aqua bidest. filter-sterilized 900.0 ml<p>Stir slowly at 4 - 10°C for 3 hours, adjust pH to 7.6 with 5 M NaOH. Filter sterilize.<p><B>Solution B:</B><br>Dissolve 14.0 g gelatine in 200 ml bidest. water, autoclave 15 minutes at 115°C.<p><B>Solution C:</B><br>Dissolve 50.05 g bovine serum albumine, fract. V (important: Sigma No A9647) in 143 ml bidest. water, or use BSA-solution Sigma A7409.<p><B>Solution D:</B><br> Rabbit serum, inactivated<br> (GIBCO 037-06120M) 86.0 ml<p>Combine the warm (37°C) solutions A, B, C, D, and filter-sterilize the warm medium. Solution B can also be added after filtration.<p>Alternatively, supplement BSK-H medium (Fa. Bioconcept, Umkirch or Sigma) with solutions B and D as above and with 100 mg/l L-glutamin from a separately autoclaved stock solution. Sterilize the mixed and warm (37°C) medium by filtration.
#>   ID_reference
#> 1        24387
#> 
#> $`132598`$culture_growth_condition$culture_temp
#>    ability test_type temp temperature_range ID_reference
#> 1 positive   optimum   30        mesophilic        24387
#> 
#> 
#> $`132598`$environment_sampling_isolation_source
#> $`132598`$environment_sampling_isolation_source$origin
#>                                            sample_type sample_date
#> 1 human blood from a patient exposed to infected ticks          NA
#> 2                    patient exposed to infected ticks          NA
#>           geo_loc_name country     continent latitude longitude
#> 1        Upper Midwest     USA North America       NA        NA
#> 2 upper midwestern USA     USA North America       NA        NA
#>   enrichment_cult_name enrichment_cult_composition enrichment_cult_duration
#> 1                   NA                          NA                       NA
#> 2                   NA                          NA                       NA
#>   enrichment_cult_temp isolation_date procedure_origin ID_reference
#> 1                   NA             NA               NA        24387
#> 2                   NA             NA               NA        25032
#> 
#> 
#> $`132598`$application_interaction
#> named list()
#> 
#> $`132598`$molecular_biology
#> $`132598`$molecular_biology$sequence
#>          DB_sequence
#> 1  GenBank/EMBL/DDBJ
#> 2  GenBank/EMBL/DDBJ
#> 3  GenBank/EMBL/DDBJ
#> 4  GenBank/EMBL/DDBJ
#> 5  GenBank/EMBL/DDBJ
#> 6  GenBank/EMBL/DDBJ
#> 7  GenBank/EMBL/DDBJ
#> 8  GenBank/EMBL/DDBJ
#> 9  GenBank/EMBL/DDBJ
#> 10 GenBank/EMBL/DDBJ
#> 11 GenBank/EMBL/DDBJ
#> 12             INSDC
#>                                                      Sequence_accession_title
#> 1                                                                        uvrA
#> 2                                                                        rplB
#> 3                                                                        recG
#> 4                                                                        pyrG
#> 5                                                                        pepX
#> 6                                                                        clpX
#> 7                                                                        clpA
#> 8                                                                        nifS
#> 9                                                                        flaB
#> 10                                                                       ospC
#> 11                                                                  rrf – rrl
#> 12 Borrelia mayonii strain MN14-1420 16S ribosomal RNA gene, partial sequence
#>    seq_acc_num sequence_length NCBI_tax_ID   NCBI_Gi ID_reference
#> 1     KM877342              NA          NA        NA        25032
#> 2     KM877343              NA          NA        NA        25032
#> 3     KM877344              NA          NA        NA        25032
#> 4     KM877345              NA          NA        NA        25032
#> 5     KM877346              NA          NA        NA        25032
#> 6     KM877347              NA          NA        NA        25032
#> 7     KM877348              NA          NA        NA        25032
#> 8     KP972469              NA          NA        NA        25032
#> 9     KR154295              NA          NA        NA        25032
#> 10    KR154297              NA          NA        NA        25032
#> 11    KR154296              NA          NA        NA        25032
#> 12    KP972468            1323     1674146 927522109        24387
#> 
#> 
#> $`132598`$strain_availability
#> $`132598`$strain_availability$strain_history
#>                          history ID_reference
#> 1 <- J. Petersen, CDC; MN14-1420        24387
#> 
#> $`132598`$strain_availability$strains
#>               strain_number ID_reference
#> 1 DSM 102811, ATCC BAA-2743        24387
#> 
#> 
#> $`132598`$references
#>   ID_reference
#> 1        20215
#> 2        24387
#> 3        25032
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          reference
#> 1                                                                                                                                                                                                                                                                                      D.Gleim, M.Kracht, N.Weiss et. al.: Prokaryotic Nomenclature Up-to-date - compilation of all names of Bacteria and Archaea, validly published according to the Bacteriological Code since 1. Jan. 1980, and validly published nomenclatural changes since. 
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                              Leibniz Institut DSMZ-Deutsche Sammlung von Mikroorganismen und Zellkulturen GmbH; Curators of the DSMZ; DSM 102811
#> 3 B. S. R.-K. Pritt, L. B.,Sloan, L. M.,Schriefer, M. E.,Replogle, A. J.,Bjork, J.,Liu, G.,Kingry, L. C.,Mead, P. S.,Neitzel, D. F.,Schiffman, E.,Hoang Johnson, D. K.,Davis, J. P.,Paskewitz, S. M.,Boxrud, D.,Deedon, A.,Lee, X.,Miller, T. K.,Feist, M. A.,Steward, C. R.,Theel, E. S.,Patel, R.,Irish, C. L.,Petersen, J. M. Borrelia mayonii sp. nov., a member of the Borrelia burgdorferi sensu lato complex, detected in patients and ticks in the upper midwestern United States. IJSEM 66: 4878-4880 2016 (10.1099/ijsem.0.001445, None)
#> 
#> 
#> $`133012`
#> $`133012`$taxonomy_name
#> $`133012`$taxonomy_name$strains_tax_PNU
#>   species_epithet subspecies_epithet is_type_strain   domain     phylum   class
#> 1      wiedmannii                 NA           TRUE Bacteria Firmicutes Bacilli
#>   ordo      family status_fam    genus  status_gen             species
#> 1   NA Bacillaceae         NA Bacillus genus (AL)  Bacillus wiedmannii
#>     status_spec                   full_scientific_name ID_reference
#> 1 sp. nov. (VP) Bacillus wiedmannii Miller et al. 2016        20215
#> 
#> $`133012`$taxonomy_name$strains
#>     domain     phylum   class       ordo      family    genus
#> 1 Bacteria Firmicutes Bacilli Bacillales Bacillaceae Bacillus
#>               species species_epithet subspecies_epithet
#> 1 Bacillus wiedmannii      wiedmannii                 NA
#>                     full_scientific_name designation is_type_strain
#> 1 Bacillus wiedmannii Miller et al. 2016 FSL W8-0169           TRUE
#>   ID_reference
#> 1        24800
#> 
#> 
#> $`133012`$morphology_physiology
#> $`133012`$morphology_physiology$cell_morphology
#>   cell_len_unit cell_width_unit gram_stain cell_len cell_width cell_shape
#> 1            µm              µm   positive      2.8        1.2 rod-shaped
#>   motility flagellum_arrangement ID_reference
#> 1       NA                    NA        25027
#> 
#> $`133012`$morphology_physiology$colony_morphology
#>   colony_len_unit hemolysis_type hemolysis_ability colony_len colony_color
#> 1              NA             NA                NA         NA       creamy
#>   colony_shape incubation_period
#> 1     circular             1 day
#>                                                medium_name medium_composition
#> 1 BHI agar (brain heart infusion medium, Becton Dickinson)                 NA
#>   ID_reference
#> 1        25027
#> 
#> $`133012`$morphology_physiology$spore_formation
#>                                                                     description
#> 1 ellipsoidal, present in the centre of vegetative cells, non-swollen sporangia
#>        type ability ID_reference
#> 1 endospore    TRUE        25027
#> 
#> $`133012`$morphology_physiology$compound_production
#>   compound_name excreted ID_reference
#> 1   haemolysin        NA        24800
#> 
#> $`133012`$morphology_physiology$oxygen_tolerance
#>             oxygen_tol ID_reference
#> 1 facultative anaerobe        25027
#> 
#> $`133012`$morphology_physiology$met_util
#>    chebi_ID                metabolite_util ability        test_type
#> 1     16808          2-Dehydro-D-gluconate       - builds acid from
#> 2     17426          5-Dehydro-D-gluconate       - builds acid from
#> 3     18305                        Arbutin       + builds acid from
#> 4        NA                         Casein       +       hydrolysis
#> 5     17057                     Cellobiose       + builds acid from
#> 6     15963                     D-Adonitol       - builds acid from
#> 7     17108                    D-Arabinose       - builds acid from
#> 8     18333                     D-Arabitol       - builds acid from
#> 9     15824                     D-Fructose       + builds acid from
#> 10    28847                       D-Fucose       - builds acid from
#> 11    17634                      D-Glucose       + builds acid from
#> 12    62318                       D-Lyxose       - builds acid from
#> 13    16899                     D-Mannitol       - builds acid from
#> 14    16024                      D-Mannose       - builds acid from
#> 15    17924                     D-Sorbitol       - builds acid from
#> 16    16443                     D-Tagatose       - builds acid from
#> 17    65327                       D-Xylose       - builds acid from
#> 18    17113                     Erythritol       - builds acid from
#> 19     4853            Esculin hydrolysate       + builds acid from
#> 20    16813                     Galactitol       - builds acid from
#> 21    28260                      Galactose       - builds acid from
#> 22     5291                        Gelatin       + builds acid from
#> 23    28066                    Gentiobiose       - builds acid from
#> 24    17754                       Glycerol       - builds acid from
#> 25    28087                       Glycogen       + builds acid from
#> 26    15443                         Inulin       - builds acid from
#> 27    30849                    L-Arabinose       - builds acid from
#> 28    18403                     L-Arabitol       - builds acid from
#> 29    18287                       L-Fucose       - builds acid from
#> 30    62345                     L-Rhamnose       - builds acid from
#> 31    17266                      L-Sorbose       - builds acid from
#> 32    65328                       L-Xylose       - builds acid from
#> 33    17306                        Maltose       + builds acid from
#> 34     6731                     Melezitose       - builds acid from
#> 35    28053                      Melibiose       - builds acid from
#> 36   320061 Methyl alpha-D-glucopyranoside       - builds acid from
#> 37    43943       Methyl alpha-D-mannoside       - builds acid from
#> 38    74863   Methyl beta-D-xylopyranoside       - builds acid from
#> 39    17268                   myo-Inositol       - builds acid from
#> 40   506227            N-Acetylglucosamine       + builds acid from
#> 41    16634                      Raffinose       - builds acid from
#> 42    33942                         Ribose       + builds acid from
#> 43    17814                        Salicin       + builds acid from
#> 44    28017                         Starch       +       hydrolysis
#> 45    17992                        Sucrose       - builds acid from
#> 46    27082                      Trehalose       + builds acid from
#> 47    27897                     Tryptophan       -    energy source
#> 48    17151                        Xylitol       - builds acid from
#>    ID_reference
#> 1         25027
#> 2         25027
#> 3         25027
#> 4         25027
#> 5         25027
#> 6         25027
#> 7         25027
#> 8         25027
#> 9         25027
#> 10        25027
#> 11        25027
#> 12        25027
#> 13        25027
#> 14        25027
#> 15        25027
#> 16        25027
#> 17        25027
#> 18        25027
#> 19        25027
#> 20        25027
#> 21        25027
#> 22        25027
#> 23        25027
#> 24        25027
#> 25        25027
#> 26        25027
#> 27        25027
#> 28        25027
#> 29        25027
#> 30        25027
#> 31        25027
#> 32        25027
#> 33        25027
#> 34        25027
#> 35        25027
#> 36        25027
#> 37        25027
#> 38        25027
#> 39        25027
#> 40        25027
#> 41        25027
#> 42        25027
#> 43        25027
#> 44        25027
#> 45        25027
#> 46        25027
#> 47        25027
#> 48        25027
#> 
#> $`133012`$morphology_physiology$met_production
#>   chebi_ID metabolite_prod production excreted ID_reference
#> 1    15688         Acetoin       TRUE       NA        25027
#> 2    35581          Indole      FALSE       NA        25027
#> 
#> $`133012`$morphology_physiology$met_test
#>   chebi_ID metabolite_test voges_proskauer methyl_red indole_test citrate_test
#> 1    15688         Acetoin               +         NA        <NA>           NA
#> 2    35581          Indole            <NA>         NA           -           NA
#>   ID_reference
#> 1        25027
#> 2        25027
#> 
#> $`133012`$morphology_physiology$enzymes
#>                    enzyme activity ec_number ID_reference
#> 1    arginine dihydrolase        -   3.5.3.6        25027
#> 2             lecithinase        +      <NA>        25027
#> 3    lysine decarboxylase        -  4.1.1.18        25027
#> 4 ornithine decarboxylase        -  4.1.1.17        25027
#> 5    tryptophan deaminase        -  4.1.99.1        25027
#> 6                  urease        -   3.5.1.5        25027
#> 
#> 
#> $`133012`$culture_growth_condition
#> $`133012`$culture_growth_condition$culture_medium
#>                           medium_name medium_growth
#> 1 NUTRIENT AGAR (DSMZ Medium 1), 30°C          TRUE
#> 2                            BHI agar          TRUE
#>                                                       media_link
#> 1 https://www.dsmz.de/microorganisms/medium/pdf/DSMZ_Medium1.pdf
#> 2                                                           <NA>
#>   medium_composition ID_reference
#> 1                 NA        24800
#> 2                 NA        25027
#> 
#> $`133012`$culture_growth_condition$culture_temp
#>    ability test_type  temp temperature_range ID_reference
#> 1 positive   optimum    30        mesophilic        24800
#> 2 positive    growth  5-43              <NA>        25027
#> 3     <NA>   optimum 20-40              <NA>        25027
#> 
#> 
#> $`133012`$environment_sampling_isolation_source
#> $`133012`$environment_sampling_isolation_source$origin
#>                                                           sample_type
#> 1                  raw milk stored in a dairy powder processing plant
#> 2 silo raw milk sample collected from a dairy powder processing plant
#>   sample_date      geo_loc_name country     continent latitude longitude
#> 1          NA              <NA>     USA North America       NA        NA
#> 2          NA north-eastern USA     USA North America       NA        NA
#>   enrichment_cult_name enrichment_cult_composition enrichment_cult_duration
#> 1                   NA                          NA                       NA
#> 2                   NA                          NA                       NA
#>   enrichment_cult_temp isolation_date procedure_origin ID_reference
#> 1                   NA           <NA>               NA        24800
#> 2                   NA           2012               NA        25027
#> 
#> 
#> $`133012`$application_interaction
#> named list()
#> 
#> $`133012`$molecular_biology
#> $`133012`$molecular_biology$sequence
#>                 DB_sequence
#> 1         GenBank/EMBL/DDBJ
#> 2 GenBank Direct submission
#> 3               INSDC (WGS)
#>                                                          Sequence_accession_title
#> 1                                                                       16S rRNA 
#> 2 Bacillus wiedmannii strain FSL W8-0169 16S ribosomal RNA gene, partial sequence
#> 3 Bacillus wiedmannii strain FSL W8-0169, whole genome shotgun sequencing project
#>           seq_acc_num sequence_length NCBI_tax_ID    NCBI_Gi ID_reference
#> 1 KU198626/SRX1297474              NA          NA         NA        25027
#> 2            KU198626            1540     1890302 1001229017        24800
#> 3        LOBC00000000             104     1890302 1003812833        24800
#> 
#> $`133012`$molecular_biology$GC_content
#>   GC_content                GC_method ID_reference
#> 1       35.3        sequence analysis        24800
#> 2       35.3 genome sequence analysis        25027
#> 
#> 
#> $`133012`$strain_availability
#> $`133012`$strain_availability$strain_history
#>                                                                                            history
#> 1 <- K. Boor, Cornell Univerity <- Matthew Watterson, Cornell University, Ithaca, USA; FSL W8-0169
#>   ID_reference
#> 1        24800
#> 
#> $`133012`$strain_availability$strains
#>                       strain_number ID_reference
#> 1 DSM 102050, CIP 111364, LMG 29269        24800
#> 
#> 
#> $`133012`$references
#>   ID_reference
#> 1        20215
#> 2        24800
#> 3        25027
#>                                                                                                                                                                                                                                                                                     reference
#> 1                                 D.Gleim, M.Kracht, N.Weiss et. al.: Prokaryotic Nomenclature Up-to-date - compilation of all names of Bacteria and Archaea, validly published according to the Bacteriological Code since 1. Jan. 1980, and validly published nomenclatural changes since. 
#> 2                                                                                                                                                                         Leibniz Institut DSMZ-Deutsche Sammlung von Mikroorganismen und Zellkulturen GmbH; Curators of the DSMZ; DSM 102050
#> 3 R. A. B. Miller, S. M.,Kent, D. J.,Carroll, L. M.,Martin, N. H.,Boor, K. J.,Kovac, J. Bacillus wiedmannii sp. nov., a psychrotolerant and cytotoxic Bacillus cereus group species isolated from dairy foods and dairy environments. IJSEM 66: 4744-4753 2016 (10.1099/ijsem.0.001421, None)
#> 
#>